+++ wgpu-hal/src/metal/mod.rs
@@ -1248,8 +1248,8 @@ pub struct ComputePipeline {
 static_assertions::assert_impl_all!(ComputePipeline: Send, Sync);
 
 impl crate::DynComputePipeline for ComputePipeline {
-    fn get_subgroup_size(&self) -> Option<usize> {
-        Some(self.raw.threadExecutionWidth())
+    fn get_subgroup_size(&self) -> Option<u32> {
+        Some(self.raw.threadExecutionWidth() as u32)
     }
 }
