+++ wgpu/src/api/compute_pipeline.rs
@@ -32,7 +32,7 @@ impl ComputePipeline {
     ///
     /// Returns `None` on all platforms except for Metal, where it returns the
     /// [`threadExecutionWidth`](https://developer.apple.com/documentation/metal/mtlcomputepipelinestate/threadexecutionwidth).
-    pub fn get_subgroup_size(&self) -> Option<usize> {
+    pub fn get_subgroup_size(&self) -> Option<u32> {
         self.inner.get_subgroup_size()
     }
