+++ wgpu/src/backend/wgpu_core.rs
@@ -1785,7 +1785,7 @@ impl dispatch::ComputePipelineInterface for CoreComputePipeline {
         .into()
     }
 
-    fn get_subgroup_size(&self) -> Option<usize> {
+    fn get_subgroup_size(&self) -> Option<u32> {
         self.wgpu_compute_pipeline.get_subgroup_size()
     }
 }
